refactor: Rename `…/plugins/aws/package/compile/events/cloudWatchLog.js`
