+++ lib/plugins/aws/package/compile/events/cloud-watch-log.js
