+++ lib/plugins/index.js
@@ -47,7 +47,7 @@ module.exports = [
   require('./aws/package/compile/events/iot.js'),
   require('./aws/package/compile/events/iotFleetProvisioning.js'),
   require('./aws/package/compile/events/cloud-watch-event.js'),
-  require('./aws/package/compile/events/cloudWatchLog.js'),
+  require('./aws/package/compile/events/cloud-watch-log.js'),
   require('./aws/package/compile/events/cognitoUserPool.js'),
   require('./aws/package/compile/events/eventBridge/index.js'),
   require('./aws/package/compile/events/sqs.js'),
